+++ cmd/skywire-cli/commands/visor/info.go
@@ -2,13 +2,12 @@ package clivisor
 
 import (
 	"fmt"
-	"log"
 	"net/http"
-	"os"
 
 	"github.com/spf13/cobra"
 
 	clirpc "github.com/skycoin/skywire/cmd/skywire-cli/commands/rpc"
+	"github.com/skycoin/skywire/cmd/skywire-cli/internal"
 	"github.com/skycoin/skywire/pkg/visor/visorconfig"
 )
 
@@ -36,97 +35,128 @@ func init() {
 var pkCmd = &cobra.Command{
 	Use:   "pk",
 	Short: "Public key of the visor",
-	Run: func(_ *cobra.Command, _ []string) {
+	Run: func(cmd *cobra.Command, _ []string) {
 		if pkg {
 			path = visorconfig.Pkgpath
 		}
+		var outputPK string
 		if path != "" {
 			conf, err := visorconfig.ReadFile(path)
 			if err != nil {
-				logger.Fatal("Failed to read config:", err)
+				internal.PrintError(cmd.Flags(), fmt.Errorf("Failed to read config: %v", err))
 			}
-			fmt.Println(conf.PK.Hex())
+			outputPK = conf.PK.Hex()
 		} else {
 			client := clirpc.Client()
 			overview, err := client.Overview()
 			if err != nil {
-				logger.Fatal("Failed to connect:", err)
+				internal.PrintError(cmd.Flags(), fmt.Errorf("Failed to connect: %v", err))
 			}
-			pk = overview.PubKey.String()
+			pk = overview.PubKey.String() + "\n"
 			if web {
 				http.HandleFunc("/", srvpk)
 				logger.Info("\nServing public key " + pk + " on port " + webPort)
 				http.ListenAndServe(":"+webPort, nil) //nolint
 			}
-			fmt.Println(overview.PubKey)
+			outputPK = overview.PubKey.Hex() + "\n"
 		}
+
+		internal.PrintOutput(cmd.Flags(), outputPK, outputPK)
 	},
 }
 
 var hvpkCmd = &cobra.Command{
 	Use:   "hvpk",
 	Short: "Public key of remote hypervisor",
-	Run: func(_ *cobra.Command, _ []string) {
+	Run: func(cmd *cobra.Command, _ []string) {
+		var hypervisors string
+
 		if pkg {
 			path = visorconfig.Pkgpath
 		}
+
 		if path != "" {
 			conf, err := visorconfig.ReadFile(path)
 			if err != nil {
-				logger.Fatal("Failed to read config:", err)
+				internal.PrintError(cmd.Flags(), fmt.Errorf("Failed to read config: %v", err))
 			}
-			fmt.Println(conf.Hypervisors)
+			hypervisors = fmt.Sprintf("%v\n", conf.Hypervisors)
 		} else {
 			client := clirpc.Client()
 			overview, err := client.Overview()
 			if err != nil {
-				logger.Fatal("Failed to connect:", err)
+				internal.PrintError(cmd.Flags(), fmt.Errorf("Failed to connect: %v", err))
 			}
-			fmt.Println(overview.Hypervisors)
+			hypervisors = fmt.Sprintf("%v\n", overview.Hypervisors)
 		}
+		internal.PrintOutput(cmd.Flags(), hypervisors, hypervisors)
 	},
 }
 
 var chvpkCmd = &cobra.Command{
 	Use:   "chvpk",
 	Short: "Public key of connected hypervisors",
-	Run: func(_ *cobra.Command, _ []string) {
+	Run: func(cmd *cobra.Command, _ []string) {
 		client := clirpc.Client()
 		overview, err := client.Overview()
 		if err != nil {
-			logger.Fatal("Failed to connect:", err)
+			internal.PrintError(cmd.Flags(), fmt.Errorf("Failed to connect: %v", err))
 		}
-		fmt.Println(overview.ConnectedHypervisor)
+		internal.PrintOutput(cmd.Flags(), overview.ConnectedHypervisor, fmt.Sprintf("%v\n", overview.ConnectedHypervisor))
 	},
 }
 
 var summaryCmd = &cobra.Command{
 	Use:   "info",
 	Short: "Summary of visor info",
-	Run: func(_ *cobra.Command, _ []string) {
+	Run: func(cmd *cobra.Command, _ []string) {
 		summary, err := clirpc.Client().Summary()
 		if err != nil {
-			log.Fatal("Failed to connect:", err)
+			internal.PrintError(cmd.Flags(), fmt.Errorf("Failed to connect: %v", err))
 		}
-		msg := fmt.Sprintf(".:: Visor Summary ::.\nPublic key: %q\nSymmetric NAT: %t\nIP: %s\nDMSG Server: %q\nPing: %q\nVisor Version: %s\nSkybian Version: %s\nUptime Tracker: %s\nTime Online: %f seconds\nBuild Tag: %s\n", summary.Overview.PubKey, summary.Overview.IsSymmetricNAT, summary.Overview.LocalIP, summary.DmsgStats.ServerPK, summary.DmsgStats.RoundTrip, summary.Overview.BuildInfo.Version, summary.SkybianBuildVersion, summary.Health.ServicesHealth, summary.Uptime, summary.BuildTag)
-		if _, err := os.Stdout.Write([]byte(msg)); err != nil {
-			log.Fatal("Failed to output build info:", err)
+		msg := fmt.Sprintf(".:: Visor Summary ::.\nPublic key: %q\nSymmetric NAT: %t\nIP: %s\nDMSG Server: %q\nPing: %q\nVisor Version: %s\nSkybian Version: %s\nUptime Tracker: %s\nTime Online: %f seconds\nBuild Tag: %s\n",
+			summary.Overview.PubKey, summary.Overview.IsSymmetricNAT, summary.Overview.LocalIP, summary.DmsgStats.ServerPK, summary.DmsgStats.RoundTrip, summary.Overview.BuildInfo.Version, summary.SkybianBuildVersion,
+			summary.Health.ServicesHealth, summary.Uptime, summary.BuildTag)
+
+		outputJSON := struct {
+			PublicKey      string  `json:"public_key"`
+			IsSymmetricNAT bool    `json:"symmetric_nat"`
+			IP             string  `json:"ip"`
+			DmsgServer     string  `json:"dmsg_server"`
+			Ping           string  `json:"ping"`
+			VisorVersion   string  `json:"visor_version"`
+			SkybianVersion string  `json:"skybian_version"`
+			UptimeTracker  string  `json:"uptime_tracker"`
+			TimeOnline     float64 `json:"time_online"`
+			BuildTag       string  `json:"build_tag"`
+		}{
+			PublicKey:      summary.Overview.PubKey.String(),
+			IsSymmetricNAT: summary.Overview.IsSymmetricNAT,
+			IP:             summary.Overview.LocalIP,
+			DmsgServer:     summary.DmsgStats.ServerPK.String(),
+			Ping:           summary.DmsgStats.RoundTrip.String(),
+			VisorVersion:   summary.Overview.BuildInfo.Version,
+			SkybianVersion: summary.SkybianBuildVersion,
+			UptimeTracker:  summary.Health.ServicesHealth,
+			TimeOnline:     summary.Uptime,
+			BuildTag:       summary.BuildTag,
 		}
+		internal.PrintOutput(cmd.Flags(), outputJSON, msg)
 	},
 }
 
 var buildInfoCmd = &cobra.Command{
 	Use:   "version",
 	Short: "Version and build info",
-	Run: func(_ *cobra.Command, _ []string) {
+	Run: func(cmd *cobra.Command, _ []string) {
 		client := clirpc.Client()
 		overview, err := client.Overview()
 		if err != nil {
-			log.Fatal("Failed to connect:", err)
+			internal.PrintError(cmd.Flags(), fmt.Errorf("Failed to connect: %v", err))
 		}
-		if _, err := overview.BuildInfo.WriteTo(os.Stdout); err != nil {
-			log.Fatal("Failed to output build info:", err)
-		}
+		buildInfo := overview.BuildInfo
+		msg := fmt.Sprintf("Version %q built on %q against commit %q\n", buildInfo.Version, buildInfo.Date, buildInfo.Commit)
+		internal.PrintOutput(cmd.Flags(), buildInfo, msg)
 	},
 }