+++ cmd/skywire-cli/commands/visor/root.go
@@ -1,8 +1,9 @@
 package clivisor
 
 import (
-	"github.com/skycoin/skycoin/src/util/logging"
 	"github.com/spf13/cobra"
+
+	"github.com/skycoin/skywire-utilities/pkg/logging"
 )
 
 var logger = logging.MustGetLogger("skywire-cli")