+++ cmd/skywire-cli/commands/visor/shutdown.go
@@ -6,6 +6,7 @@ import (
 	"github.com/spf13/cobra"
 
 	clirpc "github.com/skycoin/skywire/cmd/skywire-cli/commands/rpc"
+	"github.com/skycoin/skywire/cmd/skywire-cli/internal"
 )
 
 func init() {
@@ -15,8 +16,9 @@ func init() {
 var shutdownCmd = &cobra.Command{
 	Use:   "halt",
 	Short: "Stop a running visor",
-	Run: func(_ *cobra.Command, args []string) {
+	Run: func(cmd *cobra.Command, args []string) {
 		clirpc.Client().Shutdown() //nolint
 		fmt.Println("Visor was shut down")
+		internal.PrintOutput(cmd.Flags(), "Visor was shut down", fmt.Sprintln("Visor was shut down"))
 	},
 }